power: supply: bq27z561: Wake up from suspend on I/O operations
The current logic of using cached values in suspend is broken. It
doesn't account for bq->raw_soc which may lead to OOB access in
fg_mac_read_block(). Moreover, it could lead to overdischarge
due to not reporting the actual state of charge. Remove this logic
completely and hold a wakelock instead to address the issues.

Change-Id: Ie13f377b93c4ef73dc771279636fb555c27c3cea
Signed-off-by: Alexander Winkowski <dereference23@outlook.com>